camera: Add FlushPendingRequests to result processor
Add FlushPendingRequests method to result processor interface. Add the
implementation of this method to multiple implementation classes. This
method could be invoked by the processing block which owns the result
processor to notify the result processor for flushing of its remaining
work.

Flush all depth requests properly when flush is invoked by the
framework for the rgbird capture session.

Bug: 144532870
Bug: 146057798
Test: Pixel 4 Camera Test Checklist
Change-Id: I36c7e4fdcba3af08b047fe6057ddd489bfe6561f