EmulatedCamera: Clamp U/V planes during encoding
The jpeg encoder will process the input YUV in
batches. The height can be padded at the end so
it can be aligned with the batch size. The padding
must consist of the very last line of the respective
input buffer plane. This requirement must be
fulfilled for all planes including U/V.

Test: Camera CTS
Bug: 148018763
Change-Id: I47a1e28d9a29db82a27a93b349ac03aed5aa1b3b