Import gralloc buffers before use in EmulatedCamera
From mapper documentation:

"
A raw buffer handle must not be used to access the
underlying graphics buffer. It must be imported to
create an imported handle first.
"

Without this change, attempts to lock the stream
buffers fail when using Minigbm on Gralloc4 as the
buffers were never imported and Minigbm doesn't have
some state initialized for each buffer.

This worked previously because existing Cuttlefish
gralloc didn't check or enforce this requirement.

Bug: b/146515640
Test: cts -m CtsCameraTestCases
Change-Id: I0f71e02833399af0719c88755b13069272b60431