Directly link Lyric into GCH.
When compiling GCH with the Lyric backend, do not use dlopen
to load the symbols and instead rely on direct linking.

In order to enable direct linking, this CL changes GCH libraries
from vendor_available to vendor and moves several files to
libgooglecamerahalutils to make them accessible to the Lyric
multicam session without creating a circular dependency.

Bug: 188327280
Test: camera preview on target devices
Change-Id: I43f7f675ec078003ffa78f626046082d1148e3e1