EmulatedCamera: Use Gralloc stride for RAW requests
Currently RAW capture and CFA interpolation always
assume that the input buffer stride matches with the
buffer width. The underlying HAL format doesn't include
any such requirement and different Gralloc implementations
are allowed to use any stride that is bigger than the width
and aligned to 16.
The camera must query the current RAW stride and populate
the requested buffers accordingly.

Bug: 188992806
Test: Camera CTS
Change-Id: I8ddcfffdefac54868d64be3ff8cea12e50b8ad90