Camera: Fix failing camera VTS test.
Vts test turnOnTorchWithStrengthLevel fails for devices that don't
support this feature. The assert for METHOD_NOT_SUPPORTED fails since
the return status is INVALID_OPERATION which maps to OPERATION NOT
SUPPORTED. Modified this and changed to return UNKNOWN_TRANSACTION
which maps to METHOD_NOT_SUPPORTED status.

Test: Vts test turnOnTorchWithStrengthLevel pass on Pixel 4.
Cts test Flashlight test pass on Pixel 4.

bug: 215367324
Change-Id: I2ec41e7bfefcb0d98fc7174187472d2e33f6f5f3