Interpose __loader_dlopen instead of dlopen.
The dlopen function uses the return address of the caller to determine
a linker namespace for access. To preserve this behavior, we can
interpose __loader_dlopen instead which is passed the return address
from dlopen.

Bug: 193608778
Test: Performed go/pixel-camera-test-checklist. Also laucnhed camera
with heap snapshots enabled and verified using perfetto. Also double
checked logcat to ensure libraies are loading correctly.

Change-Id: If5651afe2454f8946fd2ea759d03a88887f31275